+++ cypress/support/idapi/auth.js
@@ -1,0 +1,11 @@
+/// <reference types="cypress" />
+
+const authRedirectSignInRecentlyEmailValidated = {
+  signInStatus: 'signedInRecently',
+  emailValidated: true,
+  redirect: null,
+};
+
+module.exports = {
+  authRedirectSignInRecentlyEmailValidated,
+};
